+++ CHANGELOG.md
@@ -14,10 +14,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 #### Added
 - methods `Frame.to_tuples()` and `Frame.to_dict()` (#1400).
 - methods `Frame.head(n)` and `Frame.tail(n)` (#1307).
+- `Frame` objects are now pickle-able (#1442).
 
 #### Fixed
 - crash when an int-column row selector is applied to a Frame which already
   had another row filter applied (#1437).
+- Frame.copy() now retains the key (#1443).
 
 
 ### [v0.7.0](https://github.com/h2oai/datatable/compare/0.7.0...v0.6.0) — 2018-11-16